Feat: Add method FavIcoLnsOut to renamed DcrdcSiteConfig
In PHP-incl/site-config.php, first rename the class DcrdcSiteConfig to DcrdcPgMeta. This change better reflects the purpose of the class, which is to extract page meta data from the site database for use in constructing web pages. Next, add the method FavIcoLnsOut, which outputs all of the favico lines to the head of the HTML document representing the web page.