Feat: Adaptive list widening for Trello boards
In «AdjusterMonkey.Trello.js», enhance click-triggered widening of Trello lists with a mutation observer that detects and responds to the addition of new lists to the board. Also make the set up of list widening more robust by using the dataset interface to prevent multiple list widening click handlers from being set up on a list.