Refactor: Website styling → Design declaration mixin
First rename design declaration mixin «decl-dsgn-maybe» in «Less/Mixins/dsgn-usage.less» to «declare-design-by-option», which is more readable. Then update the name of the mixin where it is called throughout the project. In these files, also rename design rulesets that are declared by option via the mixin to be more readable.