Feat: SS → Add modifier to content grid module
Add a modifier class to «Less/Components/grid.less» for laying out a grid whose height is responsively sized to match the viewport's minus the height of the primary navigation component. Next, reduce the breakpoint at which multi-column grid cells start to vertically stack instead of horizontally pack on wider screens. Refactor the design of the module to use rule nesting and rename the class from "grid" to the more specific "content-grid." Finally, update the file header documentation to follow a hard line length limit at 80 characters and utilize the project's latest style for marking files.

Also update the name of the design declaration flag in «Less/Settings/dsgn-usage.less».